--- lib/constants/strings.dart
@@ -1,1 +1,0 @@
-const String CALL_COLLECTION = "call";
