--- lib/videocall_module/videocall_screen_test.dart
@@ -1,522 +1,0 @@
-// import 'dart:async';
-// import 'dart:ffi';
-// import 'dart:js_util';
-
-// import 'package:cloud_firestore/cloud_firestore.dart';
-// import 'package:firebase_auth/firebase_auth.dart';
-// import 'package:flutter/foundation.dart';
-// import 'package:flutter/material.dart';
-// import 'package:flutter/scheduler.dart';
-// import 'package:flutter/widgets.dart';
-// import 'package:font_awesome_flutter/font_awesome_flutter.dart';
-// import 'package:get/get.dart';
-// import 'package:telephy/model/call.dart';
-// import 'package:telephy/utils/settings.dart';
-// import 'package:telephy/videocall_module/call_method.dart';
-
-// import 'package:agora_rtc_engine/rtc_engine.dart';
-// import 'package:agora_rtc_engine/rtc_local_view.dart' as RtcLocalView;
-// import 'package:agora_rtc_engine/rtc_remote_view.dart' as RtcRemoteView;
-
-// class CallScreen extends StatefulWidget {
-//   const CallScreen({super.key, required this.call, this.role});
-
-//   final Call call;
-//   final ClientRole? role;
-
-//   @override
-//   State<CallScreen> createState() => _CallScreenState();
-// }
-
-// class _CallScreenState extends State<CallScreen> {
-//   final CallMethods callMethods = CallMethods();
-//   final FirebaseAuth _firebaseAuth = FirebaseAuth.instance;
-//   late StreamSubscription callStreamSubscription;
-
-//   final _users = <int>[];
-//   final _infoStrings = <String>[];
-//   bool muted = false;
-//   bool video_mute = false;
-//   late RtcEngine _engine;
-
-//   @override
-//   void initState() {
-//     super.initState();
-//     addPostFramCallback();
-//     initializeAgora();
-//   }
-
-//   Future<void> initializeAgora() async {
-//     if (APP_ID.isEmpty) {
-//       setState(() {
-//         _infoStrings.add(
-//           'APP_ID missing, please provide your APP_ID in settings.dart',
-//         );
-//         _infoStrings.add('Agora Engine is not starting');
-//       });
-//       return;
-//     }
-
-//     await _initAgoraRtcEngine();
-//     _addAgoraEventHandlers();
-//     VideoEncoderConfiguration configuration = VideoEncoderConfiguration();
-//     configuration.dimensions = VideoDimensions(width: 1920, height: 1080);
-//     await _engine.setVideoEncoderConfiguration(configuration);
-//     await _engine.joinChannel(null, widget.call.channelId, null, 0);
-//   }
-
-//   Future<void> _initAgoraRtcEngine() async {
-//     _engine = await RtcEngine.create(APP_ID);
-//     await _engine.enableVideo();
-//     await _engine.setChannelProfile(ChannelProfile.LiveBroadcasting);
-//     await _engine.setClientRole(widget.role!);
-//   }
-
-//   void _addAgoraEventHandlers() {
-//     _engine.setEventHandler(RtcEngineEventHandler(error: (code) {
-//       setState(() {
-//         final info = 'onError: $code';
-//         _infoStrings.add(info);
-//       });
-//     }, joinChannelSuccess: (channel, uid, elapsed) {
-//       setState(() {
-//         final info = 'onJoinChannel: $channel, uid: $uid';
-//         _infoStrings.add(info);
-//       });
-//     }, leaveChannel: (stats) {
-//       setState(() {
-//         _infoStrings.add('onLeaveChannel');
-//         _users.clear();
-//       });
-//     }, userJoined: (uid, elapsed) {
-//       setState(() {
-//         final info = 'userJoined: $uid';
-//         _infoStrings.add(info);
-//         _users.add(uid);
-//       });
-//     }, userOffline: (uid, elapsed) {
-//       setState(() {
-//         final info = 'userOffline: $uid';
-//         _infoStrings.add(info);
-//         _users.remove(uid);
-//       });
-//     }, firstRemoteVideoFrame: (uid, width, height, elapsed) {
-//       setState(() {
-//         final info = 'firstRemoteVideo: $uid ${width}x $height';
-//         _infoStrings.add(info);
-//       });
-//     }));
-//   }
-
-//   addPostFramCallback() {
-//     SchedulerBinding.instance.addPostFrameCallback((_) {
-//       callStreamSubscription = callMethods
-//           .callStream(uid: _firebaseAuth.currentUser!.uid)
-//           .listen((DocumentSnapshot ds) {
-//         switch (ds.data) {
-//           case null:
-//             //when call hanged doc delete then pop to กันไม่ให้ snapshot is null ซ้ำ then จอดำ
-//             Navigator.pop(context);
-//             break;
-
-//           default:
-//             break;
-//         }
-//         if (ds.data == null) {
-//           Navigator.pop(context);
-//         }
-//       });
-//     });
-//   }
-
-//   Widget _expandedVideoRow(List<Widget> views) {
-//     final wrappedViews = views.map<Widget>(_videoView).toList();
-//     return Expanded(
-//       child: Row(
-//         children: wrappedViews,
-//       ),
-//     );
-//   }
-
-//   List<Widget> _getRenderViews() {
-//     final List<StatefulWidget> list = [];
-//     if (widget.role == ClientRole.Broadcaster) {
-//       list.add(RtcLocalView.SurfaceView());
-//     }
-//     _users.forEach((int uid) => list.add(
-//         RtcRemoteView.SurfaceView(channelId: widget.call.channelId, uid: uid)));
-//     return list;
-//   }
-
-//   Widget _videoView(view) {
-//     return Expanded(child: Container(child: view));
-//   }
-
-//   Widget _viewRows() {
-//     final views = _getRenderViews();
-//     switch (views.length) {
-//       case 1:
-//         return Container(
-//             child: Column(
-//           children: <Widget>[_videoView(views[0])],
-//         ));
-//       case 2:
-//         return Container(
-//             child: Column(
-//           children: <Widget>[
-//             _expandedVideoRow([views[0]]),
-//             _expandedVideoRow([views[1]])
-//           ],
-//         ));
-//       case 3:
-//         return Container(
-//             child: Column(
-//           children: <Widget>[
-//             _expandedVideoRow(views.sublist(0, 2)),
-//             _expandedVideoRow(views.sublist(2, 3))
-//           ],
-//         ));
-//       case 4:
-//         return Container(
-//             child: Column(
-//           children: <Widget>[
-//             _expandedVideoRow(views.sublist(0, 2)),
-//             _expandedVideoRow(views.sublist(2, 4))
-//           ],
-//         ));
-//       default:
-//     }
-//     return Container();
-//   }
-
-//   void _onSwitchCamera() {
-//     _engine.switchCamera();
-//   }
-
-//   void _onToggleVideo() {
-//     setState(() {
-//       video_mute = !video_mute;
-//     });
-//     _engine.enableLocalVideo(video_mute);
-//   }
-
-//   void _onToggleMute() {
-//     setState(() {
-//       muted = !muted;
-//     });
-//     _engine.muteLocalAudioStream(muted);
-//   }
-
-//   Widget _telephytoolbar() {
-//     return Scaffold(
-//       backgroundColor: Colors.grey,
-//       body: SafeArea(
-//         child: Obx(() => Container(
-//               child: Stack(
-//                 children: [
-//                   // Positioned(
-//                   //   child: Container(
-//                   //     decoration: BoxDecoration(
-//                   //       image: DecorationImage(
-//                   //         image: AssetImage("assets/images/docpic.png"),
-//                   //         fit: BoxFit.cover,
-//                   //       ),
-//                   //     ),
-//                   //   ),
-//                   // ),
-//                   Positioned(
-//                     top: 690,
-//                     child: Container(
-//                       width: 393,
-//                       height: 200,
-//                       decoration: BoxDecoration(
-//                         borderRadius: BorderRadius.all(
-//                           Radius.elliptical(400, 200),
-//                         ),
-//                         color: Color.fromRGBO(221, 222, 252, 0.6),
-//                       ),
-//                     ),
-//                   ),
-//                   // Positioned(
-//                   //   bottom: 20,
-//                   //   left: 30,
-//                   //   right: 30,
-//                   //   child: Column(
-//                   //     children: [
-//                   //       Container(
-//                   //         child: Text(
-//                   //           "${controller.state.callTime.value}",
-//                   //           style: TextStyle(
-//                   //             color: Colors.white,
-//                   //             fontSize: 20,
-//                   //             fontWeight: FontWeight.normal,
-//                   //           ),
-//                   //         ),
-//                   //       ),
-//                   //     ],
-//                   //   ),
-//                   // ),
-//                   Positioned(
-//                     bottom: 36,
-//                     left: 30,
-//                     top: 675,
-//                     child: Row(
-//                       children: [
-//                         Column(
-//                           children: [
-//                             GestureDetector(
-//                               child: Container(
-//                                 width: 70,
-//                                 height: 70,
-//                                 decoration: BoxDecoration(
-//                                   borderRadius:
-//                                       BorderRadius.all(Radius.circular(100)),
-//                                   gradient: LinearGradient(
-//                                       begin: Alignment.bottomLeft,
-//                                       end: Alignment.topRight,
-//                                       colors: [
-//                                         Color(0XFFD2ACFF),
-//                                         Color(0XFFB2DDFD)
-//                                       ]),
-//                                 ),
-//                                 child: IconButton(
-//                                   onPressed: _onToggleMute,
-//                                   color: Colors.white,
-//                                   icon: muted
-//                                       ? FaIcon(FontAwesomeIcons.microphoneSlash)
-//                                       : FaIcon(FontAwesomeIcons.microphone),
-//                                 ),
-//                               ),
-//                             ),
-//                           ],
-//                         )
-//                       ],
-//                     ),
-//                   ),
-//                   Positioned(
-//                     bottom: 60,
-//                     left: (MediaQuery.of(context).size.width - 70) / 2,
-//                     child: Row(
-//                       children: [
-//                         Column(
-//                           children: [
-//                             GestureDetector(
-//                               child: Container(
-//                                 width: 70,
-//                                 height: 70,
-//                                 decoration: BoxDecoration(
-//                                     borderRadius:
-//                                         BorderRadius.all(Radius.circular(100)),
-//                                     color:
-//                                         const Color.fromRGBO(230, 71, 71, 1.0)),
-//                                 child: IconButton(
-//                                   color: Colors.white,
-//                                   icon: FaIcon(FontAwesomeIcons.phoneSlash),
-//                                   onPressed: () => callMethods.endCall(
-//                                     call: widget.call,
-//                                   ),
-//                                 ),
-//                               ),
-//                             ),
-//                           ],
-//                         )
-//                       ],
-//                     ),
-//                   ),
-//                   Positioned(
-//                     bottom: 36,
-//                     right: 30,
-//                     top: 675,
-//                     child: Row(
-//                       children: [
-//                         Column(
-//                           children: [
-//                             GestureDetector(
-//                               child: Container(
-//                                 width: 70,
-//                                 height: 70,
-//                                 decoration: BoxDecoration(
-//                                   borderRadius:
-//                                       BorderRadius.all(Radius.circular(100)),
-//                                   gradient: LinearGradient(
-//                                       begin: Alignment.bottomLeft,
-//                                       end: Alignment.topRight,
-//                                       colors: [
-//                                         Color(0XFFD2ACFF),
-//                                         Color(0XFFB2DDFD)
-//                                       ]),
-//                                 ),
-//                                 child: IconButton(
-//                                   color: Colors.white,
-//                                   icon: video_mute
-//                                       ? FaIcon(FontAwesomeIcons.videoSlash)
-//                                       : FaIcon(FontAwesomeIcons.video),
-//                                   onPressed: _onToggleVideo,
-//                                 ),
-//                               ),
-//                             ),
-//                           ],
-//                         )
-//                       ],
-//                     ),
-//                   ),
-//                 ],
-//               ),
-//             )),
-//       ),
-//     );
-//   }
-
-//   Widget _toolbar() {
-//     if (widget.role == ClientRole.Audience) return Container();
-//     return Container(
-//       alignment: Alignment.bottomCenter,
-//       padding: const EdgeInsets.symmetric(vertical: 48),
-//       child: Row(
-//         mainAxisAlignment: MainAxisAlignment.center,
-//         children: <Widget>[
-//           RawMaterialButton(
-//             onPressed: _onToggleMute,
-//             child: Icon(
-//               muted ? Icons.mic_off : Icons.mic,
-//               color: muted ? Colors.white : Colors.blueAccent,
-//               size: 20.0,
-//             ),
-//             shape: CircleBorder(),
-//             elevation: 2.0,
-//             fillColor: muted ? Colors.blueAccent : Colors.white,
-//             padding: const EdgeInsets.all(12.0),
-//           ),
-//           RawMaterialButton(
-//             onPressed: () => callMethods.endCall(
-//               call: widget.call,
-//             ),
-//             child: Icon(
-//               Icons.call_end,
-//               color: Colors.white,
-//               size: 35.0,
-//             ),
-//             shape: CircleBorder(),
-//             elevation: 2.0,
-//             fillColor: Colors.redAccent,
-//             padding: const EdgeInsets.all(15.0),
-//           ),
-//           RawMaterialButton(
-//             onPressed: _onSwitchCamera,
-//             child: Icon(
-//               Icons.switch_camera,
-//               color: Colors.blueAccent,
-//               size: 20.0,
-//             ),
-//             shape: CircleBorder(),
-//             elevation: 2.0,
-//             fillColor: Colors.white,
-//             padding: const EdgeInsets.all(12.0),
-//           )
-//         ],
-//       ),
-//     );
-//   }
-
-//   Widget _panel() {
-//     return Container(
-//       padding: const EdgeInsets.symmetric(vertical: 48),
-//       alignment: Alignment.bottomCenter,
-//       child: FractionallySizedBox(
-//         heightFactor: 0.5,
-//         child: Container(
-//           padding: const EdgeInsets.symmetric(vertical: 48),
-//           child: ListView.builder(
-//             reverse: true,
-//             itemCount: _infoStrings.length,
-//             itemBuilder: (BuildContext context, int index) {
-//               if (_infoStrings.isEmpty) {
-//                 return Text(
-//                     "null"); // return type can't be null, a widget was required
-//               }
-//               return Padding(
-//                 padding: const EdgeInsets.symmetric(
-//                   vertical: 3,
-//                   horizontal: 10,
-//                 ),
-//                 child: Row(
-//                   mainAxisSize: MainAxisSize.min,
-//                   children: [
-//                     Flexible(
-//                       child: Container(
-//                         padding: const EdgeInsets.symmetric(
-//                           vertical: 2,
-//                           horizontal: 5,
-//                         ),
-//                         decoration: BoxDecoration(
-//                           color: Colors.yellowAccent,
-//                           borderRadius: BorderRadius.circular(5),
-//                         ),
-//                         child: Text(
-//                           _infoStrings[index],
-//                           style: TextStyle(color: Colors.blueGrey),
-//                         ),
-//                       ),
-//                     )
-//                   ],
-//                 ),
-//               );
-//             },
-//           ),
-//         ),
-//       ),
-//     );
-//   }
-
-//   @override
-//   void dispose() {
-//     // clear users
-//     _users.clear();
-//     _dispose();
-//     callStreamSubscription.cancel();
-//     super.dispose();
-//   }
-
-//   Future<void> _dispose() async {
-//     // destroy sdk
-//     await _engine.leaveChannel();
-//     await _engine.destroy();
-//   }
-
-//   @override
-//   Widget build(BuildContext context) {
-//     return Scaffold(
-//       backgroundColor: Colors.black,
-//       body: Center(
-//         child: Stack(
-//           children: <Widget>[
-//             _viewRows(),
-//             _panel(),
-//             // _telephytoolbar(),
-//             _toolbar(),
-//           ],
-//         ),
-//       ),
-//       // body: Container(
-//       //   alignment: Alignment.center,
-//       //   child: Column(
-//       //     children: <Widget>[
-//       //       Text(
-//       //         "Call has been made",
-//       //       ),
-//       //       MaterialButton(
-//       //           color: Colors.red,
-//       //           child: Icon(
-//       //             Icons.call_end,
-//       //             color: Colors.white,
-//       //           ),
-//       //           onPressed: () {
-//       //             callMethods.endCall(call: widget.call);
-//       //             Navigator.pop(context);
-//       //           }),
-//       //     ],
-//       //   ),
-//       // ),
-//     );
-//   }
-// }
